refactor: move App and Restaurant modules to modules directory
