+++ epicure-server/src/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { AppModule } from './modules/app/app.module';
 import { ValidationPipe } from '@nestjs/common';
 
 async function bootstrap() {
